+++ SharePointFramework/ProjectWebParts/src/components/ProjectStatus/Sections/ListSection/index.tsx
@@ -20,15 +20,15 @@ export class ListSection extends BaseSection<
 > {
   constructor(props: IListSectionProps) {
     super(props)
-    this.state = { isLoading: true }
+    this.state = { loading: true }
   }
 
   public async componentDidMount() {
     try {
       const data = await this._fetchData()
-      this.setState({ data, isLoading: false })
+      this.setState({ data, loading: false })
     } catch (error) {
-      this.setState({ error, isLoading: false })
+      this.setState({ error, loading: false })
     }
   }
 
@@ -52,7 +52,7 @@ export class ListSection extends BaseSection<
    * Render list
    */
   private _renderList() {
-    if (this.state.isLoading || !this.state.data) {
+    if (this.state.loading || !this.state.data) {
       return null
     }
     if (this.state.error) {